+++ .gitignore
@@ -8,7 +8,7 @@ Thumbs.db
 *.so
 coverage.*
 example
-!instrumentation/gopkg.in/confluentinc/confluent-kafka-go/kafka/otelkafka/example
+!instrumentation/github.com/confluentinc/confluent-kafka-go/kafka/otelkafka/example
 
 instrumentation/google.golang.org/grpc/otelgrpc/example/server/server
 instrumentation/google.golang.org/grpc/otelgrpc/example/client/client
