+++ instrumentation/github.com/confluentinc/confluent-kafka-go/kafka/otelkafka/consumer.go
@@ -17,7 +17,7 @@ package otelkafka
 import (
 	"fmt"
 
-	"gopkg.in/confluentinc/confluent-kafka-go.v1/kafka"
+	"github.com/confluentinc/confluent-kafka-go/kafka"
 
 	oteltrace "go.opentelemetry.io/otel/api/trace"
 	"go.opentelemetry.io/otel/label"
