+++ instrumentation/github.com/confluentinc/confluent-kafka-go/kafka/otelkafka/producer.go
@@ -19,7 +19,7 @@ package otelkafka
 import (
 	"fmt"
 
-	"gopkg.in/confluentinc/confluent-kafka-go.v1/kafka"
+	"github.com/confluentinc/confluent-kafka-go/kafka"
 
 	oteltrace "go.opentelemetry.io/otel/api/trace"
 	"go.opentelemetry.io/otel/codes"
